backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic>=2.7.4
openai>=1.109.1
langchain>=1.0.0
langchain-openai>=1.0.0
langchain-community>=0.4.0
langchain-chroma>=1.0.0
chromadb>=1.0.0
pypdf>=3.17.0
python-docx>=1.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
tiktoken>=0.7.0
